backend/requirements.txt
Flask>=3.0,<4.0
flask-cors>=4.0,<5.0
tensorflow>=2.13,<3.0
numpy>=1.26,<3.0
Pillow>=10.0,<12.0
